@@ -1,7 +1,7 @@
 {
   "name": "@weni/unnnic-system",
-  "version": "3.2.7",
-  "type": "commonjs",
+  "version": "3.2.8",
+  "type": "module",
   "files": [
     "dist",
     "src",
@@ -30,7 +30,7 @@
     "build": "vite build",
     "build:types": "vue-tsc && vite build",
     "preview": "vite preview",
-    "lint": "eslint --ext .js,.vue,.ts --ignore-path .gitignore --fix src",
+    "lint": "eslint src/**/*.{js,vue,ts} --fix",
     "format": "prettier --write src/",
     "type-check": "vue-tsc --noEmit",
     "storybook": "storybook dev -p 6006",
@@ -80,14 +80,15 @@
     "@vitest/coverage-istanbul": "^1.6.0",
     "@vitest/coverage-v8": "^1.6.0",
     "@vitest/ui": "^1.6.0",
-    "@vue/eslint-config-prettier": "^9.0.0",
+    "@vue/eslint-config-prettier": "^10.2.0",
+    "@vue/eslint-config-typescript": "^14.6.0",
     "@vue/test-utils": "^2.4.6",
-    "eslint": "^8.0.0",
-    "eslint-config-prettier": "^9.1.0",
-    "eslint-plugin-prettier": "^5.1.3",
+    "@weni/eslint-config": "^2.0.0",
+    "eslint": "^9.33.0",
     "eslint-plugin-storybook": "^0.8.0",
-    "eslint-plugin-vue": "^9.25.0",
+    "eslint-plugin-vue": "10.4.0",
     "jsdom": "^24.1.1",
+    "prettier": "3.6.2",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "sass": "^1.63.0",
