@@ -170,7 +170,7 @@ watch(
   () => {
     validateProps();
   },
-  { deep: true, immediate: true }
+  { deep: true, immediate: true },
 );
 </script>
 
