@@ -1,95 +1,95 @@
-import type { Component } from "vue";
-import { unnnicFontSize as fontSize } from "./config";
-import formElement from "./FormElement/FormElement.vue";
-import input from "./Input/Input.vue";
-import inputNext from "./InputNext/InputNext.vue";
-import inputDatePicker from "./InputDatePicker/InputDatePicker.vue";
-import button from "./Button/Button.vue";
-import buttonIcon from "./Button/ButtonIcon.vue";
-import sidebar from "./Sidebar/index.vue";
-import sidebarItem from "./Sidebar/SidebarItem.vue";
-import table from "./Table/Table.vue";
-import tableRow from "./Table/TableRow.vue";
-import dropdown from "./Dropdown/Dropdown.vue";
-import dropdownItem from "./Dropdown/DropdownItem.vue";
-import avatarIcon from "./AvatarIcon/AvatarIcon.vue";
-import icon from "./Icon.vue";
-import iconLoading from "./IconLoading/IconLoading.vue";
-import toolTip from "./ToolTip/ToolTip.vue";
-import card from "./Card/Card.vue";
-import cardSimple from "./Card/SimpleCard.vue";
-import cardCompany from "./Card/CardCompany.vue";
-import cardData from "./Card/CardData.vue";
-import cardImage from "./CardImage/CardImage.vue";
+import type { Component } from 'vue';
+import { unnnicFontSize as fontSize } from './config';
+import formElement from './FormElement/FormElement.vue';
+import input from './Input/Input.vue';
+import inputNext from './InputNext/InputNext.vue';
+import inputDatePicker from './InputDatePicker/InputDatePicker.vue';
+import button from './Button/Button.vue';
+import buttonIcon from './Button/ButtonIcon.vue';
+import sidebar from './Sidebar/index.vue';
+import sidebarItem from './Sidebar/SidebarItem.vue';
+import table from './Table/Table.vue';
+import tableRow from './Table/TableRow.vue';
+import dropdown from './Dropdown/Dropdown.vue';
+import dropdownItem from './Dropdown/DropdownItem.vue';
+import avatarIcon from './AvatarIcon/AvatarIcon.vue';
+import icon from './Icon.vue';
+import iconLoading from './IconLoading/IconLoading.vue';
+import toolTip from './ToolTip/ToolTip.vue';
+import card from './Card/Card.vue';
+import cardSimple from './Card/SimpleCard.vue';
+import cardCompany from './Card/CardCompany.vue';
+import cardData from './Card/CardData.vue';
+import cardImage from './CardImage/CardImage.vue';
 // import cardFlow from './CardFlow/CardFlow.vue';
-import cardProject from "./CardProject/CardProject.vue";
-import cardInformation from "./CardInformation/CardInformation.vue";
-import checkbox from "./Checkbox/Checkbox.vue";
-import collapse from "./Collapse/Collapse.vue";
-import radio from "./Radio/Radio.vue";
-import languageSelect from "./Dropdown/LanguageSelect.vue";
-import modal from "./Modal/Modal.vue";
-import modalUpload from "./ModalUpload/ModalUpload.vue";
-import { callAlert, callModal } from "../utils/call";
-import selectSmart from "./SelectSmart/SelectSmart.vue";
+import cardProject from './CardProject/CardProject.vue';
+import cardInformation from './CardInformation/CardInformation.vue';
+import checkbox from './Checkbox/Checkbox.vue';
+import collapse from './Collapse/Collapse.vue';
+import radio from './Radio/Radio.vue';
+import languageSelect from './Dropdown/LanguageSelect.vue';
+import modal from './Modal/Modal.vue';
+import modalUpload from './ModalUpload/ModalUpload.vue';
+import { callAlert, callModal } from '../utils/call';
+import selectSmart from './SelectSmart/SelectSmart.vue';
 // import select from './Select/Select.vue';
-import selectItem from "./Select/SelectItem.vue";
+import selectItem from './Select/SelectItem.vue';
 // import selectListItem from './SelectListItem/SelectListItem.vue';
-import multiSelect from "./MultiSelect/MultiSelect.vue";
-import alert from "./Alert/Alert.vue";
+import multiSelect from './MultiSelect/MultiSelect.vue';
+import alert from './Alert/Alert.vue';
 // import autocomplete from './Input/Autocomplete.vue';
 // import autocompleteSelect from './AutocompleteSelect/AutocompleteSelect.vue';
-import tag from "./Tag/Tag.vue";
-import accordion from "./Accordion/Accordion.vue";
-import indicator from "./Indicator/Indicator.vue";
-import skeletonLoading from "./SkeletonLoading/SkeletonLoading.vue"; // TODO: Unbreak this component
-import carousel from "./Carousel/Carousel.vue";
-import label from "./Label/Label.vue";
-import tab from "./Tab/Tab.vue";
-import tabsExpanded from "./TabsExpanded/TabsExpanded.vue";
-import banner from "./Banner/Banner.vue";
-import comment from "./Comment/Comment.vue";
-import datePicker from "./DatePicker/DatePicker.vue";
-import breadcrumb from "./Breadcrumb/Breadcrumb.vue";
-import Switch from "./Switch/Switch.vue";
-import Slider from "./Slider/Slider.vue";
-import DataArea from "./DataArea/DataArea.vue";
-import Pagination from "./Pagination/Pagination.vue";
-import DropArea from "./DropArea/DropArea.vue";
-import UploadArea from "./UploadArea/UploadArea.vue";
-import ImportCard from "./ImportCard/ImportCard.vue";
-import DateFilter from "./DateFilter/DateFilter.vue";
-import ChatText from "./ChatText/ChatText.vue";
-import TextArea from "./TextArea/TextArea.vue";
-import CardNumber from "./CardNumber/CardNumber.vue";
-import chartRainbow from "./ChartRainbow/ChartRainbow.vue";
-import chartBar from "./ChartBar/ChartBar.vue";
-import chartLine from "./ChartLine/ChartLine.vue";
-import moodRating from "./MoodRating/MoodRating.vue";
-import starRating from "./StarRating/StarRating.vue";
-import audioRecorder from "./AudioRecorder/AudioRecorder.vue";
-import circleProgressBar from "./CircleProgressBar/CircleProgressBar.vue";
-import progressBar from "./ProgressBar/ProgressBar.vue";
-import ChatsContact from "./ChatsContact/ChatsContact.vue";
-import ChatsDashboardTagLive from "./ChatsDashboardTagLive/ChatsDashboardTagLive.vue";
-import ChatsHeader from "./ChatsHeader/ChatsHeader.vue";
-import ChatsMessage from "./ChatsMessage/ChatsMessage.vue";
-import ReplyMessage from "./ChatsMessage/ReplyMessage.vue";
-import ChatsNavbar from "./ChatsNavbar/ChatsNavbar.vue";
-import ChatsUserAvatar from "./ChatsUserAvatar/ChatsUserAvatar.vue";
-import ChartMultiLine from "./ChartMultiLine/ChartMultiLine.vue";
-import EmojiPicker from "./EmojiPicker/EmojiPicker.vue";
-import ChartFunnel from "./ChartFunnel/ChartFunnel.vue";
-import Disclaimer from "./Disclaimer/Disclaimer.vue";
-import Drawer from "./Drawer/Drawer.vue";
-import TableNext from "./TableNext/TableNext.vue";
-import ModalNext from "./ModalNext/ModalNext.vue";
-import ModalDialog from "./ModalDialog/ModalDialog.vue";
-import Tour from "./Tour/Tour.vue";
-import Navigator from "./Navigator/index.vue";
-import SelectTime from "./SelectTime/index.vue";
-import DataTable from "./DataTable/index.vue";
-import Chip from "./Chip/Chip.vue";
+import tag from './Tag/Tag.vue';
+import accordion from './Accordion/Accordion.vue';
+import indicator from './Indicator/Indicator.vue';
+import skeletonLoading from './SkeletonLoading/SkeletonLoading.vue'; // TODO: Unbreak this component
+import carousel from './Carousel/Carousel.vue';
+import label from './Label/Label.vue';
+import tab from './Tab/Tab.vue';
+import tabsExpanded from './TabsExpanded/TabsExpanded.vue';
+import banner from './Banner/Banner.vue';
+import comment from './Comment/Comment.vue';
+import datePicker from './DatePicker/DatePicker.vue';
+import breadcrumb from './Breadcrumb/Breadcrumb.vue';
+import Switch from './Switch/Switch.vue';
+import Slider from './Slider/Slider.vue';
+import DataArea from './DataArea/DataArea.vue';
+import Pagination from './Pagination/Pagination.vue';
+import DropArea from './DropArea/DropArea.vue';
+import UploadArea from './UploadArea/UploadArea.vue';
+import ImportCard from './ImportCard/ImportCard.vue';
+import DateFilter from './DateFilter/DateFilter.vue';
+import ChatText from './ChatText/ChatText.vue';
+import TextArea from './TextArea/TextArea.vue';
+import CardNumber from './CardNumber/CardNumber.vue';
+import chartRainbow from './ChartRainbow/ChartRainbow.vue';
+import chartBar from './ChartBar/ChartBar.vue';
+import chartLine from './ChartLine/ChartLine.vue';
+import moodRating from './MoodRating/MoodRating.vue';
+import starRating from './StarRating/StarRating.vue';
+import audioRecorder from './AudioRecorder/AudioRecorder.vue';
+import circleProgressBar from './CircleProgressBar/CircleProgressBar.vue';
+import progressBar from './ProgressBar/ProgressBar.vue';
+import ChatsContact from './ChatsContact/ChatsContact.vue';
+import ChatsDashboardTagLive from './ChatsDashboardTagLive/ChatsDashboardTagLive.vue';
+import ChatsHeader from './ChatsHeader/ChatsHeader.vue';
+import ChatsMessage from './ChatsMessage/ChatsMessage.vue';
+import ReplyMessage from './ChatsMessage/ReplyMessage.vue';
+import ChatsNavbar from './ChatsNavbar/ChatsNavbar.vue';
+import ChatsUserAvatar from './ChatsUserAvatar/ChatsUserAvatar.vue';
+import ChartMultiLine from './ChartMultiLine/ChartMultiLine.vue';
+import EmojiPicker from './EmojiPicker/EmojiPicker.vue';
+import ChartFunnel from './ChartFunnel/ChartFunnel.vue';
+import Disclaimer from './Disclaimer/Disclaimer.vue';
+import Drawer from './Drawer/Drawer.vue';
+import TableNext from './TableNext/TableNext.vue';
+import ModalNext from './ModalNext/ModalNext.vue';
+import ModalDialog from './ModalDialog/ModalDialog.vue';
+import Tour from './Tour/Tour.vue';
+import Navigator from './Navigator/index.vue';
+import SelectTime from './SelectTime/index.vue';
+import DataTable from './DataTable/index.vue';
+import Chip from './Chip/Chip.vue';
 
 type VueComponent = Component;
 
