@@ -1,6 +1,6 @@
-import type { App, Component } from "vue";
+import type { App, Component } from 'vue';
 
-export type { IconProps } from "../components/Icon.vue";
+export type { IconProps } from '../components/Icon.vue';
 
 export interface UnnnicPlugin {
   install(app: App): void;
