@@ -1,4 +1,4 @@
-export type SchemeColor = 
+export type SchemeColor =
   | 'background-solo'
   | 'background-sky'
   | 'background-grass'
@@ -99,4 +99,3 @@ export type SchemeColor =
   | 'teal-800'
   | 'teal-900'
   | 'teal-950';
-  
