@@ -2,4 +2,4 @@ declare module '*.vue' {
   import type { DefineComponent } from 'vue';
   const component: DefineComponent<{}, {}, any>;
   export default component;
-} 
+}
