@@ -2,6 +2,6 @@ declare module '@vueuse/components' {
   export interface OnClickOutsideOptions {
     [key: string]: any;
   }
-  
+
   export const vOnClickOutside: any;
-} 
+}
